fix int -> i64

+++ uniform_linear_sushi/answer-ugis/main.cpp
@@ -13,15 +13,15 @@ int main(){
     vector<pair<i64,i64>> sushi(N);
 
     for(int i = 0; i < N; i++){
-        int S, T;
+        i64 S, T;
         cin >> S >> T;
         sushi[i] = {S+T, S};
     }
 
     sort(sushi.begin(), sushi.end());
-    int t = 0, ans = 0;
+    i64 t = 0, ans = 0;
     for(auto& p : sushi){
-        int s, e;
+        i64 s, e;
         tie(e, s) = p;
         if(t <= s) {
             t = e;
